// src/data/gymSpritesGenerated.ts
export interface GymTrainerSprite {
  graphicsId: string;
  spriteSheet: string;
  spriteWidth: number;
  spriteHeight: number;
  spriteFrame: number;
}
// ...
export const GYM_LEADER_EXTRA_SPRITES: Partial<Record<string, GymTrainerSprite[]>> = {
  "TRAINER_TATE_AND_LIZA_1": [{
    graphicsId: "OBJ_EVENT_GFX_LIZA",
    spriteSheet: "sprites/trainers/liza.png",
    spriteWidth: 16,
    spriteHeight: 32,
    spriteFrame: 0,
  }],
};